LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-fit/hittest-expected.txt

123
4567

FAIL Text selection by a pointer device should work correctly assert_equals: expected "67" but got ""
